login to e2e tests with explicit email
