+++ ui_tests/caseworker/features/case_officer.feature
@@ -5,7 +5,7 @@ Feature: I want to assign a dedicated case officer / single point of contact to 
 
   @skip @legacy
   Scenario: Assign and remove case officer
-    Given I sign in to SSO or am signed into SSO
+    Given I sign in as "test-uat-user@digital.trade.gov.uk" # /PS-IGNORE
     And I create open application or open application has been previously created
     When I go to application previously created
     And I click Assign Case Officer Button
