+++ ui_tests/caseworker/features/give_advice/lu_nlr_advice.feature
@@ -7,7 +7,7 @@ Feature: I want to record my user advice and any comments and conditions relatin
   @lu_nlr_advice
   Scenario: LU NLR advice journey
     # Setup
-    Given I sign in to SSO or am signed into SSO
+    Given I sign in as "test-uat-user@digital.trade.gov.uk" # /PS-IGNORE
     And I create standard application or standard application has been previously created
     And I prepare the application for final review NLR
     # Scenario starts
