+++ ui_tests/caseworker/features/give_advice/mod_approve_advice.feature
@@ -7,7 +7,7 @@ Feature: I want to record my user advice and any comments and conditions relatin
   @mod_approve_advice
   Scenario: MOD approve advice journey
     ##### MOD to circulate a case #####
-    Given I sign in to SSO or am signed into SSO
+    Given I sign in as "test-uat-user@digital.trade.gov.uk" # /PS-IGNORE
     And I create standard application or standard application has been previously created
     # LR
     When I switch to "Licensing Reception" with queue "Licensing Reception SIEL applications" and I submit the case
