+++ ui_tests/caseworker/features/view_cases_queue_view.feature
@@ -5,7 +5,7 @@ Feature: I want to view cases in the queue view
   So that I can make review the case before making any decisions
 
   Scenario: Gov user can see case in queue view
-    Given I sign in to SSO or am signed into SSO
+    Given I sign in as "test-uat-user@digital.trade.gov.uk" # /PS-IGNORE
     And I create an application with <name>,<product>,<part_number>,<clc_rating>,<end_user_name>,<end_user_address>,<consignee_name>,<consignee_address>,<country>,<end_use>
     And the status is set to "submitted"
     When I go to the case list page
@@ -16,7 +16,7 @@ Feature: I want to view cases in the queue view
     | Test    | Rifle   | PN-123/XYZ  | PL9002      | Automated End user | 1234, High street | BE      | Automated Consignee | 1234, Trade centre  | Research and development |
 
   Scenario: Gov user can customise queue view
-    Given I sign in to SSO or am signed into SSO
+    Given I sign in as "test-uat-user@digital.trade.gov.uk" # /PS-IGNORE
     And I create an application with <name>,<product>,<part_number>,<clc_rating>,<end_user_name>,<end_user_address>,<consignee_name>,<consignee_address>,<country>,<end_use>
     And the status is set to "submitted"
     When I go to the case list page
