+++ ui_tests/caseworker/features/view_goods.feature
@@ -2,7 +2,7 @@
 Feature: I want to be able to view goods on a case
 
   Scenario: View goods detail page
-    Given I sign in to SSO or am signed into SSO
+    Given I sign in as "test-uat-user@digital.trade.gov.uk" # /PS-IGNORE
     And I create standard application or standard application has been previously created
     When I go to application previously created
     And I click the first good on the case
