+++ ui_tests/caseworker/step_defs/test_give_advice/test_lu_change_licence_status.py
@@ -1,29 +1,8 @@
 from pytest_bdd import scenarios, when, then, parsers
 from selenium.webdriver.common.by import By
 from ui_tests.caseworker.pages.shared import Shared
-from ui_tests.caseworker.pages.mock_signin_page import MockSigninPage
-from django.conf import settings
-from selenium.common.exceptions import NoSuchElementException
-from django.conf import settings
 
 scenarios("../../features/give_advice/lu_licence_change_status.feature", strict_gherkin=False)
-
-
-@when("I logout")  # noqa
-def i_logout(driver, internal_url):  # noqa
-    driver.get(internal_url.rstrip("/") + "/auth/logout/")
-
-
-@when(parsers.parse('I sign in as "{email}"'))
-def mock_sso_caseworker_sign_in_again(driver, internal_url, email):  # noqa
-    driver.get(internal_url)
-    try:
-        mock_sso_login_screen = driver.find_element(By.XPATH, "//*[contains(text(), 'Mock SSO Login')]")
-    except NoSuchElementException:
-        mock_sso_login_screen = None
-
-    if mock_sso_login_screen and settings.MOCK_SSO_ACTIVATE_ENDPOINTS:
-        MockSigninPage(driver).sign_in(email)
 
 
 @when("I click change licence status")
